(chore): sets cluster role binding as owner
cluster role binding as ownerRef on managed service account
